Update server structure to match auth&user
